# improve ListDirectories performance