--- a/Godeps/Godeps.json
+++ b/Godeps/Godeps.json
@@ -488,6 +488,10 @@
 			"Rev": "0b12d6b521d83fc7f755e7cfc1b1fbdd35a01a74"
 		},
 		{
+			"ImportPath": "github.com/karrick/godirwalk",
+			"Rev": "2de2192f9e35ce981c152a873ed943b93b79ced4"
+		},
+		{
 			"ImportPath": "github.com/klauspost/crc32",
 			"Rev": "a3b15ae34567abb20a22992b989cd76f48d09c47"
 		},