--- a/container/common/helpers_test.go
+++ b/container/common/helpers_test.go
@@ -1,0 +1,28 @@
+// Copyright 2018 Google Inc. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package common
+
+import (
+	"testing"
+)
+
+func BenchmarkListDirectories(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		output := make(map[string]struct{})
+		if err := ListDirectories("/sys/fs/cgroup", "", true, output); err != nil {
+			b.Fatal(err)
+		}
+	}
+}