… vendor/github.com/karrick/godirwalk/.gitignore
… vendor/github.com/karrick/godirwalk/LICENSE
… vendor/github.com/karrick/godirwalk/README.md
… vendor/github.com/karrick/godirwalk/dirent.go
… vendor/github.com/karrick/godirwalk/doc.go
… vendor/github.com/karrick/godirwalk/examples/clean-empties/main.go
… vendor/github.com/karrick/godirwalk/examples/walk-fast/main.go
… vendor/github.com/karrick/godirwalk/examples/walk-stdlib/main.go
… vendor/github.com/karrick/godirwalk/go.mod
… vendor/github.com/karrick/godirwalk/go.sum
… vendor/github.com/karrick/godirwalk/readdir.go
… vendor/github.com/karrick/godirwalk/readdir_test.go
… vendor/github.com/karrick/godirwalk/readdir_unix.go
… vendor/github.com/karrick/godirwalk/readdir_windows.go
… vendor/github.com/karrick/godirwalk/testdata/dir1/dir1a/file1a1
… vendor/github.com/karrick/godirwalk/testdata/dir1/dir1a/skip
… vendor/github.com/karrick/godirwalk/testdata/dir1/dir1a/z1a2
… vendor/github.com/karrick/godirwalk/testdata/dir1/file1b
… vendor/github.com/karrick/godirwalk/testdata/dir2/file2a
… vendor/github.com/karrick/godirwalk/testdata/dir2/skip/file2b1
… vendor/github.com/karrick/godirwalk/testdata/dir2/z2c/file2c1
… vendor/github.com/karrick/godirwalk/testdata/dir3/aaa.txt
… vendor/github.com/karrick/godirwalk/testdata/dir3/skip
… vendor/github.com/karrick/godirwalk/testdata/dir3/zzz/aaa.txt
… vendor/github.com/karrick/godirwalk/testdata/dir4/aaa.txt
… vendor/github.com/karrick/godirwalk/testdata/dir4/symlinkToDirectory
… vendor/github.com/karrick/godirwalk/testdata/dir4/symlinkToFile
… vendor/github.com/karrick/godirwalk/testdata/dir4/zzz/aaa.txt
… vendor/github.com/karrick/godirwalk/testdata/dir5/a1.txt
… vendor/github.com/karrick/godirwalk/testdata/dir5/a2/a2a/a2a1.txt