used github.com/karrick/godirwalk to reduces syscall.
```
name               old time/op    new time/op    delta
ListDirectories-2    17.4ms ± 9%     3.7ms ±15%  -78.88%  (p=0.000 n=9+9)

name               old alloc/op   new alloc/op   delta
ListDirectories-2    3.55MB ± 0%    0.48MB ± 0%  -86.36%  (p=0.000 n=9+10)

name               old allocs/op  new allocs/op  delta
ListDirectories-2     20.8k ± 0%     11.6k ± 0%  -44.13%  (p=0.000 n=10+10)
```