@dashpole 2018-10-29
`The following files are not properly formatted: container/common/endian_little.go container/common/helpers_unix.go container/common/endian_little.go container/common/helpers_unix.go`