@dashpole 2018-10-29
Is there a reason why you added a new repo and vendored it in here?  That might make it harder for us to make changes in the future.  Are there any other more commonly used implementations of this?

I'm also not sure if we can use BSD licenses or not.  Let me check...