@orisano 2018-10-29
for licensing problem. my code based on golang/go private functions.
golang/go license is [here](https://github.com/golang/go/blob/master/LICENSE).

When I try to specify license on code, boilerplate check fails.
what will i do?