@dashpole 2018-10-29
Is there a golang function we can use that isn't private?  Yeah, we keep all of the code in the repo that we don't own.  If it is code from golang, then the proper thing to do is keep it under their license and vendor it in.  If you can actually add a commit which is the exact copy of the original golang code, and then your seperate changes on top of that, that would be easier for me to review.  Also link to the original golang code if possible.