@orisano 2018-10-29
## References
https://github.com/golang/go/blob/master/src/syscall/syscall_linux.go#L862-L880
https://github.com/golang/go/blob/master/src/os/dir_unix.go#L47-L88
https://github.com/golang/go/blob/master/src/syscall/dirent.go
https://github.com/golang/go/blob/master/src/syscall/endian_big.go
https://github.com/golang/go/blob/master/src/syscall/endian_little.go