@dashpole 2018-11-01
So the proper way to do this is:

- make your own fork of the original go repo (the entire thing)
- make you changes to that fork
- vendor that fork into cAdvisor

You leave their license and boilerplate (header of the files), just as it is.
I'm not sure how licenses work if you are just copying files, so please use the above approach.