@orisano 2018-11-01
to used github.com/karrick/godirwalk to resolve license problem.