@sashankreddya 2018-11-07
@orisano / @dashpole : Looks like after this commit I am unable to recognize "mesos" containers in the directory path "/sys/fs/cgroup/<cgroup-name>/mesos". I have verified without including this commit things work fine. 
I am debugging it, you guys could share any inputs you have.