@orisano 2018-11-07
Is using symlink in "/sys/fs/cgroup/mesos" ?
currently, maybe unhandle symlink type.

by the way, Is "/sys/fs/cgroup//mesos" equals "/sys/fs/cgroup/mesos" ?