@sashankreddya 2018-11-07
no symlinks are used and yes it should be the cgroups path is "/sys/fs/cgroup/<cpu,memory>/mesos"
I could see the error thrown by 
```
dirents, err := godirwalk.ReadDirents(dirpath, buf)
```
in listDirectories when "updateSubcontainers" tries to fetch the subcontainers.


Error
```
error(*github.com/google/cadvisor/vendor/github.com/pkg/errors.withStack) *{
	error: error(*github.com/google/cadvisor/vendor/github.com/pkg/errors.withMessage) *{
		cause: error(*os.PathError) ...,
		msg: "cannot Open",},
	stack: *github.com/google/cadvisor/vendor/github.com/pkg/errors.stack len: 15, cap: 32, [11136028,11135472,11164142,11163830,11167113,18958881,19150681,19113714,19174192,19183722,19178323,19514576,19512011,19484179,4585041],}
(dlv) n
```